retirement_tracker/requirements.txt
django~=3.2.5
python-dateutil~=2.8.2